+++ hub/integrations/__init__.py
@@ -1,3 +1,3 @@
 from .pytorch import dataset_to_pytorch
-from .skorch import pytorch_module_to_skorch
+from .skorch import skorch
 from .tf import dataset_to_tensorflow
