+++ hub/integrations/skorch/__init__.py
@@ -1,1 +1,1 @@
-from .skorch import pytorch_module_to_skorch
+from .skorch import skorch
